requirements.txt
# Server requirements
fastapi==0.116.2
uvicorn==0.35.0
pydantic==2.11.9
python-dotenv==1.1.1
httpx==0.28.1
redis==6.4.0
prometheus-client==0.22.1  # For metrics
tenacity==9.1.2  # For retries
structlog==25.4.0  # For structured logging
diffusers==0.35.1
apscheduler==3.11.0
asyncpg==0.30.0  # For PostgreSQL async support
pyjwt[crypto]==2.10.1  # Required for JWT authentication with RSA support
cryptography==45.0.7  # For secure token handling
faiss-cpu==1.12.0
aiohttp==3.12.15
numpy==2.3.3
sentence-transformers==5.1.0
cohere==5.18.0
transformers==4.56.1
python-multipart==0.0.20
psutil==7.0.0
watchfiles==1.1.0
websockets==15.0.1
grpcio==1.78.0  # For gRPC communication

# Protocol Buffers for gRPC
protobuf==7.34.0